Removing the "Sandia" tag that was added to the Team page
Rational: there is currently no tag for any other organization, not even
a CEA one for the CExA Tea-Time page, and that "Sandia" tag is
exclusively used here to annotate the Team page.
It looks just out of place and yield a "Team" "Sandia" marking which is
not the message we are trying to convey.